host: stop ddt filling threads when upload completes
On fast enough serial connections, it is possible for the upload to
complete before the DDT fill threads finish.  Without notifying the
fill threads that their service is no longer needed, it is possible
that the entire upload process may have to wait for minutes before
the fill threads terminate.