provide an alternative way of running host code
Instead of relying on tip(1) or other tool like it to redirect stdin and
stdout to the host binary, we can no use host-raw which takes an extra
argument (the serial device) and reads and writes directly to it.  This new
host binary does *not* redirect user's stdin to the serial port.  It simply
waits for a special string (!!!) to appear on the serial port to initiate
the transfer of files.  Before the transfer begins and after it ends, it
duplicates all received data on stdout.